pyproject.toml
[project]
name = "ai-cover-letter"
version = "0.1.0"
description = "AI-powered cover letter generator with document parsing and LinkedIn integration"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[all]>=0.104.0",
    "uvicorn>=0.24.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.12.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "selenium>=4.15.0",
    "webdriver-manager>=4.0.0",
    "duckduckgo-search>=4.1.0",
    "python-docx>=1.1.0",
    "PyPDF2>=3.0.0",
    "openpyxl>=3.1.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "python-dateutil>=2.8.0",
    "pytz>=2023.3",
    "jinja2>=3.1.0",
    "markdown>=3.5.0",
    "ipykernel>=6.29.5",
    "docling>=2.39.0",
    "sentence-transformers>=5.0.0",
    "torch>=2.7.1",
    "reportlab>=4.4.2",
    "google-generativeai>=0.8.5",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "black==23.11.0",
    "isort==5.12.0",
    "flake8==6.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"] 
